Save preferences to QSettings
Remove the old xml preferences file and replace it with
QSettings instead. Also divide all of the settings into
logical groups.

Task-number: QT3DS-777

+++ src/Authoring/Qt3DStudio/UI/RecentItems.h
@@ -53,8 +53,8 @@ public:
     void AddRecentItem(const QString &inItem);
     void RemoveRecentItem(const QString &inItem, bool rebuild = true);
 
-    QString GetItem(long inIndex) const;
-    long GetItemCount() const { return (long)m_RecentItems.size(); }
+    QString GetItem(size_t inIndex) const;
+    size_t GetItemCount() const { return m_RecentItems.size(); }
 
 protected:
     void ReconstructList();
